backend/requirements.txt
fastapi[standard]
uvicorn
markitdown[pdf, docx, pptx]
openai
python-dotenv
